% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyloadermessage.R
\name{withLoaderMessage}
\alias{withLoaderMessage}
\title{Add a loader message to a UI element}
\usage{
withLoaderMessage(ui_element, message = "Loading...")
}
\arguments{
\item{ui_element}{UI element}

\item{message}{message}
}
\description{
Add a loader message to a UI element
}
